requirements.txt
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.12.3
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
et_xmlfile==2.0.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
greenlet==3.2.4
gspread==6.2.1
gspread-dataframe==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.3.4
oauthlib==3.3.1
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.3.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-telegram-bot==22.5
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
selenium==4.21.0
shiboken6==6.10.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wsproto==1.2.0
